Integration of RabiitMq communication

+++ ReactiveXComponent/Common/Header.cs
@@ -6,7 +6,7 @@ namespace ReactiveXComponent.Common
 
         public long ComponentCode { get; set; }
 
-        public long EventCode { get; set; }
+        public int EventCode { get; set; }
 
         public string MessageType { get; set; }
